I've added detection for 5 different sections of the paddle.  Each section is 7 pixels long (the ball is 6 so there is some wiggle room).  Based upon where the ball hits it will return in different directions and different speeds.  Feels a little janky though and I need to tune it some.
Need to work on scoring next so that the ball will disappear after someone misses it and then starts going towards the alternate team member.  Also need to play with paddle location, might need to bring them in a little bit. Also need to work on the computer AI.  Works great if ball is going up and down but no good if it's going just left to right.