Added the font and the scores are now on both sides of the screen.  Combined balldir and ballspeed into a single variable.  Using percentages of the paddle size to determine angle of reflect as well as speed.
Need to add no person playing mode, update so that there is a pause between scoring and the ball going into play again, updating the game to end at a certain score, and having the ball serve to the loser of the last round.  Misc tuning will be needed as well.